move initMeters func to test utils, make use of it and check events in unit test
